fix: use platform-safe temporary directory instead of hardcoded /tmp
- Replace unsafe '/tmp' fallback with path_provider's getTemporaryDirectory()
- Ensures cross-platform compatibility for Android, iOS, desktop platforms
- Addresses high-priority review comment about mobile platform safety

Fixes: https://github.com/ahmet-cetinkaya/whph/pull/43#discussion_r2310311677
